repo: backport git-bridge actions from ZMS5
