+++ Products/zms/ZMSRepositoryManager.py
@@ -32,16 +32,9 @@ from Products.zms import IZMSRepositoryManager
 from Products.zms import IZMSRepositoryProvider
 from Products.zms import ZMSItem
 from Products.zms import _fileutil
+from Products.zms import _repositoryutil
 from Products.zms import standard
 from Products.zms import zopeutil
-
-
-def get_class(py):
-  if isinstance(py,bytes):
-    py = py.decode('utf-8')
-  id = re.findall('class (.*?):', py)[0]
-  exec(py)
-  return eval(id)
 
 
 ################################################################################
@@ -318,116 +311,14 @@ class ZMSRepositoryManager(
 
     def remoteFiles(self, provider):
       standard.writeLog(self,"[remoteFiles]: provider=%s"%str(provider))
-      r = {}
       basepath = self.get_conf_basepath(provider.id)
-      if os.path.exists(basepath):
-        def traverse(base, path):
-          names = os.listdir(path)
-          for name in names:
-            filepath = os.path.join(path, name)
-            if os.path.isdir(filepath):
-              traverse(base,filepath)
-            elif name.startswith('__') and name.endswith('__.py'):
-              # Read python-representation of repository-object
-              standard.writeLog(self,"[remoteFiles]: read %s"%filepath)
-              f = open(filepath,"rb")
-              py = f.read()
-              f.close()
-              # Analyze python-representation of repository-object
-              d = {}
-              try:
-                c = get_class(py)
-                d = c.__dict__
-              except:
-                d['revision'] = standard.writeError(self,"[traverse]: can't analyze filepath=%s"%filepath)
-              id = d.get('id',name)
-              rd = {}
-              rd['id'] = id
-              rd['filename'] = filepath[len(base)+1:]
-              rd['data'] = py
-              rd['version'] = d.get("revision",self.getLangFmtDate(os.path.getmtime(filepath),'eng'))
-              r[rd['filename']] = rd
-              # Read artefacts and avoid processing of hidden files, e.g. .DS_Store on macOS 
-              for file in [x for x in names if x != name and not x.startswith('.')]:
-                artefact = os.path.join(path,file)
-                if os.path.isfile(artefact):
-                  standard.writeLog(self,"[remoteFiles]: read artefact %s"%artefact)
-                  f = open(artefact,"rb")
-                  data = f.read()
-                  f.close()
-                  rd = {}
-                  rd['id'] = id
-                  rd['filename'] = artefact[len(base)+1:]
-                  rd['data'] = data
-                  rd['version'] = self.getLangFmtDate(os.path.getmtime(artefact),'eng')
-                  r[rd['filename']] = rd
-        traverse(basepath,basepath)
-      return r
+      return _repositoryutil.remoteFiles(self, basepath)
 
 
     def readRepository(self, provider):
       standard.writeLog(self,"[readRepository]: provider=%s"%str(provider))
-      r = {}
       basepath = self.get_conf_basepath(provider.id)
-      if os.path.exists(basepath):
-        def traverse(base, path):
-          names = os.listdir(path)
-          for name in names:
-            filepath = os.path.join(path, name)
-            if os.path.isdir(filepath):
-              traverse(base, filepath)
-            elif name.startswith('__') and name.endswith('__.py'):
-              # Read python-representation of repository-object
-              standard.writeLog(self,"[readRepository]: read %s"%filepath)
-              f = open(filepath, "rb")
-              py = f.read()
-              f.close()
-              try:
-                if isinstance(py, bytes):
-                  py = py.decode('utf-8')
-              except:
-                pass
-              # Analyze python-representation of repository-object
-              d = {}
-              try:
-                c = get_class(py)
-                d = c.__dict__
-              except:
-                d['revision'] = standard.writeError(self,"[readRepository]: ")
-              id = d.get('id',name)
-              r[id] = {}
-              for k in [x for x in d if not x.startswith('__')]:
-                v = d[k]
-                if inspect.isclass(v):
-                  dd = v.__dict__
-                  v = []
-                  for kk in [x for x in dd if not x.startswith('__')]:
-                    vv = dd[kk]
-                    # Try to read artefact.
-                    if 'id' in vv:
-                      fileprefix = vv['id'].split('/')[-1]
-                      for file in [x for x in names if x==fileprefix or x.startswith('%s.'%fileprefix)]:
-                        artefact = os.path.join(path, file)
-                        standard.writeLog(self,"[readRepository]: read artefact %s"%artefact)
-                        f = open(artefact, "rb")
-                        data = f.read()
-                        f.close()
-                        try:
-                          if isinstance(data, bytes):
-                            data = data.decode('utf-8')
-                        except:
-                          pass
-                        vv['data'] = data
-                        break
-                    if isinstance(py,bytes):
-                      v.append((py.decode('utf-8').find('\t\t%s ='%kk), vv))
-                    else:
-                      v.append((py.find('\t\t%s ='%kk), vv))
-                  v.sort()
-                  v = [x[1] for x in v]
-                r[id][k] = v
-        traverse(basepath, basepath)
-      return r
+      return _repositoryutil.readRepository(self, basepath)
 
 
     """
@@ -474,7 +365,11 @@ class ZMSRepositoryManager(
                 f = open(filepath,"wb")
                 if isinstance(data,str):
                   data = data.encode("utf-8")
-                f.write(data)
+                try:
+                  f.write(data)
+                except:
+                  # f.write('## ERROR ON FILE WRITE')
+                  f.write(data.encode("utf-8"))
                 f.close()
               else:
                 failure.append('%s is None'%file)
